+++ CHANGELOG.md
@@ -6,6 +6,7 @@ https://keepachangelog.com/en/1.0.0/
 - `SynchronousTaskWait`: Rewritten to use `IOperation`
 - `SynchronousTaskWait`: Preserves leading trivia when applying the code fix
 - `SynchronousTaskWait`: Doesn't offer a code fix when a timeout is passed to `Task.Wait(timeoutMs)`
+- `ComparingStringsWithoutStringComparison`: Preserves leading trivia when applying the code fix
 
 ## [1.17.1] - 2023-01-02
 - `ElementaryMethodsOfTypeInCollectionNotOverridden`: Fixed an issue where a `NullReferenceException` would be thrown